+++ mcio_remote/instance.py
@@ -230,9 +230,6 @@ class Launcher:
         env["MCIO_MODE"] = self.run_options.mcio_mode
         env["MCIO_ACTION_PORT"] = str(self.run_options.action_port)
         env["MCIO_OBSERVATION_PORT"] = str(self.run_options.observation_port)
-        env["MCIO_FRAME_TYPE"] = str(self.run_options.frame_type)
-        if self.run_options.frame_type == "JPEG":
-            env["MCIO_FRAME_QUALITY"] = str(self.run_options.frame_quality)
         env["MCIO_HIDE_WINDOW"] = str(self.run_options.hide_window)
         return env